Refactored the Document_Library folder and incoporated flow from DL into Reader
